restDJ_Murat/requirements.txt
Django>=4.2
djangorestframework>=3.15
asgiref>=3.8
pytz>=2024.2
sqlparse>=0.5.0

